+++ library/src/main/java/com/google/android/exoplayer/extractor/webm/VarintReader.java
@@ -1,5 +1,6 @@
 package com.google.android.exoplayer.extractor.webm;
 
+import com.google.android.exoplayer.C;
 import com.google.android.exoplayer.extractor.ExtractorInput;
 
 import java.io.EOFException;
@@ -19,8 +20,8 @@ import java.io.IOException;
    *
    * <p>{@code 0x80} is a one-byte integer, {@code 0x40} is two bytes, and so on up to eight bytes.
    */
-  private static final int[] VARINT_LENGTH_MASKS = new int[] {
-    0x80, 0x40, 0x20, 0x10, 0x08, 0x04, 0x02, 0x01
+  private static final long[] VARINT_LENGTH_MASKS = new long[] {
+    0x80L, 0x40L, 0x20L, 0x10L, 0x08L, 0x04L, 0x02L, 0x01L
   };
 
   private final byte[] scratch;
@@ -53,48 +54,41 @@ import java.io.IOException;
    *
    * @param input The {@link ExtractorInput} from which the integer should be read.
    * @param allowEndOfInput True if encountering the end of the input having read no data is
-   *     allowed, and should result in {@code -1} being returned. False if it should be
-   *     considered an error, causing an {@link EOFException} to be thrown.
-   * @param removeLengthMask Removes the variable-length integer length mask from the value
-   * @return The read value, or -1 if {@code allowEndOfStream} is true and the end of the input was
-   *     encountered.
+   *     allowed, and should result in {@link C#RESULT_END_OF_INPUT} being returned. False if it
+   *     should be considered an error, causing an {@link EOFException} to be thrown.
+   * @param removeLengthMask Removes the variable-length integer length mask from the value.
+   * @param maximumAllowedLength Maximum allowed length of the variable integer to be read.
+   * @return The read value, or {@link C#RESULT_END_OF_INPUT} if {@code allowEndOfStream} is true
+   *     and the end of the input was encountered, or {@link C#RESULT_MAX_LENGTH_EXCEEDED} if the
+   *     length of the varint exceeded maximumAllowedLength.
    * @throws IOException If an error occurs reading from the input.
    * @throws InterruptedException If the thread is interrupted.
    */
   public long readUnsignedVarint(ExtractorInput input, boolean allowEndOfInput,
-      boolean removeLengthMask) throws IOException, InterruptedException {
+      boolean removeLengthMask, int maximumAllowedLength) throws IOException, InterruptedException {
     if (state == STATE_BEGIN_READING) {
       // Read the first byte to establish the length.
       if (!input.readFully(scratch, 0, 1, allowEndOfInput)) {
-        return -1;
+        return C.RESULT_END_OF_INPUT;
       }
       int firstByte = scratch[0] & 0xFF;
-      length = -1;
-      for (int i = 0; i < VARINT_LENGTH_MASKS.length; i++) {
-        if ((VARINT_LENGTH_MASKS[i] & firstByte) != 0) {
-          length = i + 1;
-          break;
-        }
-      }
+      length = parseUnsignedVarintLength(firstByte);
       if (length == -1) {
         throw new IllegalStateException("No valid varint length mask found");
       }
       state = STATE_READ_CONTENTS;
     }
 
+    if (length > maximumAllowedLength) {
+      state = STATE_BEGIN_READING;
+      return C.RESULT_MAX_LENGTH_EXCEEDED;
+    }
+
     // Read the remaining bytes.
     input.readFully(scratch, 1, length - 1);
 
-    // Parse the value.
-    if (removeLengthMask) {
-      scratch[0] &= ~VARINT_LENGTH_MASKS[length - 1];
-    }
-    long varint = 0;
-    for (int i = 0; i < length; i++) {
-      varint = (varint << 8) | (scratch[i] & 0xFF);
-    }
     state = STATE_BEGIN_READING;
-    return varint;
+    return assembleVarint(scratch, length, removeLengthMask);
   }
 
   /**
@@ -102,6 +96,43 @@ import java.io.IOException;
    */
   public int getLastLength() {
     return length;
+  }
+
+  /**
+   * Parses and the length of the varint given the first byte.
+   *
+   * @param firstByte First byte of the varint.
+   * @return Length of the varint beginning with the given byte if it was valid, -1 otherwise.
+   */
+  public static int parseUnsignedVarintLength(int firstByte) {
+    int varIntLength = -1;
+    for (int i = 0; i < VARINT_LENGTH_MASKS.length; i++) {
+      if ((VARINT_LENGTH_MASKS[i] & firstByte) != 0) {
+        varIntLength = i + 1;
+        break;
+      }
+    }
+    return varIntLength;
+  }
+
+  /**
+   * Assemble a varint from the given byte array.
+   *
+   * @param varintBytes Bytes that make up the varint.
+   * @param varintLength Length of the varint to assemble.
+   * @param removeLengthMask Removes the variable-length integer length mask from the value.
+   * @return Parsed and assembled varint.
+   */
+  public static long assembleVarint(byte[] varintBytes, int varintLength,
+      boolean removeLengthMask) {
+    long varint = varintBytes[0] & 0xFFL;
+    if (removeLengthMask) {
+      varint &= ~VARINT_LENGTH_MASKS[varintLength - 1];
+    }
+    for (int i = 1; i < varintLength; i++) {
+      varint = (varint << 8) | (varintBytes[i] & 0xFFL);
+    }
+    return varint;
   }
 
 }