+++ RELEASENOTES.md
@@ -1,8 +1,18 @@
 # Release notes #
 
-### Current dev branch (from r1.5.0) ###
+### Current dev branch (from r1.5.1) ###
 
+* [Nothing yet]
+
+### r1.5.0 ###
+
+* Enable smooth frame release by default.
 * Added OkHttpDataSource extension.
+* AndroidTV: Correctly detect 4K display size on Bravia devices.
+* FMP4: Handle non-sample data in mdat boxes.
+* TTML: Fix parsing of some colors on Jellybean.
+* SmoothStreaming: Ignore tfdt boxes.
+* Misc bug fixes.
 
 ### r1.5.0 ###